requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary>=2.9.11
wheel
aio-pika
pydantic-settings
